// plugins/Pageflow/ui/core/types.ts
export type Errors = Record<string, string>
export type ErrorBag = Record<string, Errors>
// ...
export type RequestPayload = Record<string, FormDataConvertible> | FormData

export interface PageProps {
  [key: string]: unknown
}

export interface Page<SharedProps extends PageProps = PageProps> {
  component: string
  props: PageProps &
    SharedProps & {
      errors: Errors & ErrorBag
      deferred?: Record<string, VisitOptions['only']>
    }
  url: string
  version: string | null
  clearHistory: boolean
  encryptHistory: boolean
  deferredProps?: Record<string, VisitOptions['only']>
  mergeProps?: string[]
  deepMergeProps?: string[]
  matchPropsOn?: string[]

  /** @internal */
  rememberedState: Record<string, unknown>
}
// ...
export type Visit<T extends RequestPayload = RequestPayload> = {
  method: Method
  data: T
  replace: boolean
  preserveScroll: PreserveStateOption
  preserveState: PreserveStateOption
  only: Array<string>
  except: Array<string>
  headers: Record<string, string>
  errorBag: string | null
  forceFormData: boolean
  queryStringArrayFormat: 'indices' | 'brackets'
  async: boolean
  showProgress: boolean
  prefetch: boolean
  fresh: boolean
  reset: string[]
  preserveUrl: boolean
  loadPage: boolean
}
// ...
export type VisitCallbacks = {
  onCancelToken: { ({ cancel }: { cancel: VoidFunction }): void }
  onBefore: GlobalEventCallback<'before'>
  onStart: GlobalEventCallback<'start'>
  onProgress: GlobalEventCallback<'progress'>
  onFinish: GlobalEventCallback<'finish'>
  onCancel: GlobalEventCallback<'cancel'>
  onSuccess: GlobalEventCallback<'success'>
  onError: GlobalEventCallback<'error'>
  onPrefetched: GlobalEventCallback<'prefetched'>
  onPrefetching: GlobalEventCallback<'prefetching'>
}

export type VisitOptions<T extends RequestPayload = RequestPayload> = Partial<Visit<T> & VisitCallbacks>
// ...
export type PendingVisitOptions = {
  url: URL
  completed: boolean
  cancelled: boolean
  interrupted: boolean
}

export type PendingVisit = Visit & PendingVisitOptions

export type ActiveVisit = PendingVisit & Required<VisitOptions>